# Docs/openapi.yaml
openapi: 3.0.3
info:
  title: Oasis API
  description: API routes for Oasis parametric insurance platform
  version: 1.0.0
servers:
  - url: https://your-app.vercel.app
    description: Production
  - url: http://localhost:3000
    description: Local development
paths:
...
  /api/webhooks/disruption:
    post:
      tags:
        - Webhooks
      summary: Realtime disruption webhook
      description: |
        Accepts a single disruption trigger from providers that support push (e.g. Tomorrow.io Alerts).
        Runs processSingleTrigger immediately: inserts event, finds eligible policies, fraud checks, creates claims and payouts.
        Secured with WEBHOOK_SECRET or CRON_SECRET (Authorization Bearer or X-Webhook-Secret).
      operationId: postWebhooksDisruption
      security:
        - webhookSecret: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [type, lat, lng]
              properties:
                type: { type: string, enum: [weather, traffic, social] }
                subtype: { type: string, example: "extreme_heat" }
                severity: { type: number, minimum: 0, maximum: 10, default: 7 }
                lat: { type: number }
                lng: { type: number }
                radius_km: { type: number, default: 15 }
                raw: { type: object }
      responses:
        '200':
          description: Processed
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  eventId: { type: string, format: uuid }
                  claimsCreated: { type: integer }
                  payoutsInitiated: { type: integer }
        '400':
          description: Invalid body
        '401':
          description: Unauthorized
        '503':
          description: Webhook or Supabase not configured
...
  /api/onboarding/verify-government-id:
    post:
      tags:
        - KYC & Onboarding
      summary: Verify government ID
      description: |
        Uploads and validates standard Indian government identification using Vision LLMs.

        The model cross-references the `full_name` provided with the legible text on the `aadhaar` card.
      operationId: postVerifyGovernmentId
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file, full_name, id_type]
              properties:
                file: { type: string, format: binary }
                full_name: { type: string, example: "Rahul Sharma" }
                id_type: { type: string, enum: [aadhaar], example: "aadhaar" }
      responses:
        '200':
          description: Verification result
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean, example: true }
                  verified: { type: boolean, example: true }
                  government_id_url: { type: string, example: "https://bucket..." }
                  reason: { type: string, example: "Aadhaar verified and name matched." }
  /api/payments/create-checkout:
    post:
      tags:
        - Payments & Subscriptions
      summary: Create Stripe Checkout
      description: Initiates a synchronous Stripe Checkout session for activating a weekly delivery risk premium.
      operationId: postCreateCheckout
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [planId, weekStart, weekEnd]
              properties:
                planId: { type: string, format: uuid, example: "basic-plan-uuid" }
                weekStart: { type: string, format: date, example: "2026-03-09" }
                weekEnd: { type: string, format: date, example: "2026-03-15" }
      responses:
        '200':
          description: Redirect URL to Stripe hosted checkout page.
          content:
            application/json:
              schema:
                type: object
                properties:
                  url: { type: string, example: "https://checkout.stripe.com/c/pay/..." }
                  sessionId: { type: string, example: "cs_test_..." }
                  policyId: { type: string, format: uuid, example: "policy-uuid..." }
...
components:
  securitySchemes:
    cronSecret:
      type: apiKey
      in: header
      name: Authorization
      description: Bearer CRON_SECRET
    webhookSecret:
      type: apiKey
      in: header
      name: Authorization
      description: Bearer WEBHOOK_SECRET or CRON_SECRET (or X-Webhook-Secret)